git-gui: fix detection of Cygwin
MSys2 might *look* like Cygwin, but it is *not* Cygwin... Unless it
is run with `MSYSTEM=MSYS`, that is.

This change is backwards-compatible because MSys also used the
MSYSTEM variable to determine whether to run in MSys or in MinGW
mode. The only difference is that MSys does not support running
in 64-bit mode at all.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>